Move SchemaXmlParser and PropertiesXmlParser to Admin bundle and make XmlParserTrait internal
